fix #21

+++ include/tween.h
@@ -548,7 +548,7 @@ namespace tweeny {
              * @returns current values
              * @sa seek
              */
-            const typename detail::tweentraits<T, Ts...>::valuesType & jump(int32_t point, bool suppressCallbacks = false);
+            const typename detail::tweentraits<T, Ts...>::valuesType & jump(size_t point, bool suppressCallbacks = false);
 
             /**
              * @brief Returns the current tween point
@@ -624,7 +624,7 @@ namespace tweeny {
             tween<T> & forward(); ///< @sa tween::forward
             tween<T> & backward(); ///< @sa tween::backward
             int direction() const; ///< @sa tween::direction
-            const T & jump(int32_t point, bool suppressCallbacks = false); ///< @sa tween::jump
+            const T & jump(size_t point, bool suppressCallbacks = false); ///< @sa tween::jump
             uint16_t point() const; ///< @sa tween::point
 
         private /* member types */:
